requirements.txt
aniso8601==10.0.0
attrs==24.3.0
bcpy==0.1.8
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-restx==1.3.0
greenlet==3.1.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.36.1
rpds-py==0.22.3
sentry-sdk==2.20.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
